NetworkSecurityProject/requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
mlflow
pyaml
dagshub==0.3.23
fastapi
uvicorn
python-multipart


##-e .